More Work on Volume Rendering
Testing experimental Python code for todays meeting. Installed necessary modules to run the code. Will attempt to run the Python code in Unity as a solution to the volume rendering issue. Did not push resulting .txt and .png files from running the Python code because they were over 100MB in size and could not be pushed to GitHub.com.