Progress on Xray Simulation (from team meeting)
Worked on "xraySimulation.cs" that is equivalent "xrayGenerator.py", but in c sharp. The file is capable of capturing xray images of a .nii file at a specified angle. We still want to create support for moving the table up/down or left/right, etc.. Meeting again this Friday to attempt to deploy the code and (assuming we are successful) potentially add more support for different c-arm movements.